Move mocks to own folder and namespace
